feat: add logo.svg to resources/images
This commit adds the company logo in SVG format to the
`resources/images/` directory. The logo is a vector graphic,
ensuring scalability and visual clarity across different
resolutions and devices.

feat: add new SVG graphics with gradient fills

This commit adds new SVG graphics to the project. The graphics
utilize linear gradients with a variety of color stops to create
visually appealing elements. The paths are defined with specific
coordinates and transformations to achieve the desired shapes and
effects. These graphics can be used for decorative purposes or to
enhance the user interface.

```diff
--- a/src/assets/images/background.svg
+++ b/src/assets/images/background.svg
@@ -1,5 +1,5 @@
-<!-- Generator: Adobe Illustrator 25.2.0, SVG Export Plug-In . SVG Version: 6.00 Build 0) -->
-<svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
+<!-- Generator: Adobe Illustrator 25.2.0, SVG Export Plug-In . SVG Version: 6.00 Build 0) -->
+<svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
 viewBox="0 0 500 500" style="enable-background:new 0 0 500 500;" xml:space="preserve">
 <style type="text/css">
 .st0{fill:#29D3DA;}
@@ -10,10 +10,10 @@
 .st4{fill:#961484;}
 </style>
 <linearGradient id="SVGID_1_" gradientUnits="userSpaceOnUse" x1="39.383" y1="120.2424" x2="47.247" y2="106.8724" gradientTransform="matrix(7.8769 0 0 -7.8769 364.1079 4969.5996)">
- <stop offset="0" style="stop-color:#29D3DA"/>
- <stop offset="0.519" style="stop-color:#0077FF"/>
- <stop offset="0

feat: add new SVG graphics with gradient fills

This commit adds a set of new SVG graphics to the project.
The graphics are defined using path elements with gradient
fills, utilizing linear gradients with various color stops.
The gradients are defined with unique IDs and applied to the
paths using the `fill:url()` style attribute. These new
graphics enhance the visual appeal of the application.

```diff
--- a/public/assets/images/logo.svg
+++ b/public/assets/images/logo.svg
@@ -14,6 +14,7 @@
 <g>
 <defs>
 <rect id="SVGID_1_" width="512" height="512"/>
+
 </defs>
 <clipPath id="SVGID_2_">
 <use xlink:href="#SVGID_1_" overflow="visible"/>
@@ -21,6 +22,7 @@
 <g clip-path="url(#SVGID_2_)">
 <linearGradient id="SVGID_3_" gradientUnits="userSpaceOnUse" x1="238.7305" y1="44.8813" x2="273.2695" y2="44.8813">
 <stop offset="0" style="stop-color:#29D3DA"/>
+
 <stop offset="0.519" style="stop-color:#0077FF"/>
 <stop offset="0.999" style="stop-color:#064093"/>
 <stop offset="1" style="stop-color:#084698"/>
@@ -28,6 +30,7 @@
 <path style="fill:url(#SVGID_3_);" d="M256,12.5c93.053,0,168.5,75.447,168.5,168.5s-75.447,168.5-168.5,168.5
 S87.5,281.553,87.5,188.5S162.947,12.5,256,12.5 M256,0C114.609,0,0,114.609,0,256s114.

```
feat: add new logo

This commit adds a new logo to the project. The logo is an SVG file
with multiple gradients and paths.
```

feat: add new logo

This commit adds a new logo in SVG format. The logo consists of
various shapes and gradients, creating a visually appealing design.
The logo is intended to improve the branding and visual identity
of the application.

feat(app-layout.tsx): replace text logo with image logo in AppLayout
feat(auth-layout.tsx): replace icon logo with image logo in AuthLayout
feat(app-layout.tsx): make profile and settings labels to be links
feat(app-layout.tsx): import logo and profile routes